+++ scalyr_agent/builtin_monitors/syslog_monitor.py
@@ -627,6 +627,9 @@ class SyslogRequestParser(object):
                 limit_key="syslog-no-frames",
             )
 
+        self._remaining = self._remaining[self._offset :]
+        self._offset = 0
+
 
 class SyslogRawRequestParser(SyslogRequestParser):
     """
@@ -709,6 +712,8 @@ class SyslogBatchedRequestParser(SyslogRequestParser):
 
         while self._offset < size:
             # 2->TODO use slicing to get bytes in both python versions.
+            # TODO: This is not really robust, we should make it an explicit config option if
+            # we should try to parse messages as framed or new line delimited one.
             c = self._remaining[self._offset : self._offset + 1]
             framed = b"0" <= c <= b"9"
 
@@ -802,7 +807,7 @@ class SyslogTCPHandler(six.moves.socketserver.BaseRequestHandler):
     # each new connection.
 
     def __init__(self, *args, **kwargs):
-        self.request_parser = kwargs.pop("request_parser", False)
+        self.request_parser = kwargs.pop("request_parser", "default")
         self.incomplete_frame_timeout = kwargs.pop("incomplete_frame_timeout", None)
         self.message_delimiter = kwargs.pop("message_delimiter", "\n")